Reorganize validate tests
